+++ server.js
@@ -18,8 +18,8 @@ const sessionMiddleware = session({
     resave: false, 
     saveUninitialized: false, 
     cookie:{
-        secure:false, // Only set to true if you are using HTTPS.
-        httpOnly:false, // Only set to true if you are using HTTPS.
+        secure: process.env.HTTPS || false, // Only set to true if you are using HTTPS.
+        httpOnly: process.env.HTTPS || false, // Only set to true if you are using HTTPS.
         maxAge:3600000 * 8 // Session max age in milliseconds. 3600000 = 1 hour.
     }
 });